==> notebooks/06_Manipulating_DataFrames_to_build_one_single_DataFrame.ipynb <==
In this notebook, I aim to load the data saved from previous scraping notebooks and explore a way to pivot it in order to construct a final, unified DataFrame. This DataFrame will contain transplant data across multiple years for all transplant centers in Italy.

import pandas as pd 

# import one file
df1= pd.read_csv('../data_raw/2024/2024_Cuore.csv')
df1.head()

# removing the 'Totale Trapianti' column
df1.drop(labels='Totale Trapianti', axis=1, inplace=True)
df1.tail()

# removing the last row with 'Totale' in the 'Struttura trapianto' column
df1.drop(labels=df1.index[-1], inplace=True)
df1.tail()

# pivot the DataFrame from wide to long format (the desired format)
df1 = pd.melt(df1, id_vars= ['Struttura trapianto'], var_name='Organo', value_name='Numero')
    
# Adding Year to the dataframe
df1['Anno'] = 2024
df1.head()

# Load the CSV file for 2024 - Cuore
df = pd.read_csv('../data_raw/2024/2024_Rene.csv')

# Drop the 'Totale Trapianti' column if it exists
if 'Totale Trapianti' in df.columns:
    df.drop(columns='Totale Trapianti', inplace=True)

# Drop the last row if it contains the total summary
if df.iloc[-1, 0].strip().lower() in ['totale', 'totali']:
    df.drop(index=df.index[-1], inplace=True)

df['Città'] = df['Struttura trapianto'].str.split('-').str[0].str.strip()

# Reshape the DataFrame from wide to long format
df_long = pd.melt(
    df,
    id_vars=['Struttura trapianto', 'Città'],
    var_name='Sottotipo',
    value_name='Numero'
)

# Add metadata columns
df_long['Organo'] = 'Rene'
df_long['Anno'] = 2024

# Optional: Reorder columns
df_long = df_long[['Struttura trapianto','Città', 'Organo', 'Sottotipo', 'Numero', 'Anno']]

# Preview
df_long.head()


df_long['Sottotipo'].unique()
